stereo_camera/requirements.txt
opencv-contrib-python>=4.5.0
numpy>=1.21.0
PyYAML>=5.4.0
